app/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dataclasses-json==0.6.7
dnspython==2.7.0
email_validator==2.2.0
fastapi-cli==0.0.7
fastapi-server-session==0.0.1
fastapi==0.115.11
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx-sse==0.4.0
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain-community==0.3.20
langchain-core==0.3.47
langchain-ollama==0.3.0
langchain-postgres==0.0.13
langchain-text-splitters==0.3.7
langchain==0.3.21
langgraph-checkpoint-postgres==2.0.19
langgraph-checkpoint==2.0.21
langgraph-prebuilt==0.1.4
langgraph-sdk==0.1.58
langgraph==0.3.18
langsmith==0.3.18
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
msgpack==1.1.0
multidict==6.2.0
mypy-extensions==1.0.0
numpy==2.2.4
ollama==0.4.7
orjson==3.10.15
packaging==24.2
pgvector==0.3.6
propcache==0.3.0
psycopg-binary==3.2.6
psycopg-pool==3.2.6
psycopg==3.2.6
pycparser==2.22
pydantic-settings==2.8.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests-toolbelt==1.0.0
requests==2.32.3
rich-toolkit==0.13.2
rich==13.9.4
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
tavily-python==0.5.1
tenacity==9.0.0
tiktoken==0.9.0
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
yarl==1.18.3
zstandard==0.23.0
